feat(suite): open suite desktop from troubleshooting tip directly
